// src/styling/types.ts
import { ColorValue } from 'react-native';

export type Theme = {
  color: {
    background: ColorValue;
    blue1: ColorValue;
    blue2: ColorValue;
    blue3: ColorValue;
    gray1: ColorValue;
    gray2: ColorValue;
    label: ColorValue;
    orange: ColorValue;
    placeholder: ColorValue;
    secondaryBackground: ColorValue;
    secondaryLabel: ColorValue;
    separator: ColorValue;
    yellow: ColorValue;
  };
};